requirements.txt

python-dotenv==1.0.1
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1

pdfminer.six==20231228
python-docx==1.1.2
python-pptx==0.6.23
markdown2==2.4.13

Pillow==10.4.0
pytesseract==0.3.13

faster-whisper==1.0.3
ffmpeg-python==0.2.0
pydub==0.25.1
youtube-transcript-api==0.6.2
yt-dlp==2024.8.6

google-generativeai==0.7.2
numpy==1.26.4
tqdm==4.66.5
